backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
werkzeug==3.0.1
pypdf2==3.0.1
python-pptx==0.6.23
langchain==0.1.13
langchain-google-genai==0.0.11
langchain-community==0.0.29
faiss-cpu==1.8.0
python-dotenv==1.0.1
